move Arel::Subquery extention into MultiTenant, fix issues for Rails < 4.2
